xsk: add umem to netdev in best-effort mode, when zero-copy fails
Commit c9b47cc1fabc ("xsk: fix bug when trying to use both copy and
zero-copy on one queue id") stores the umem into the netdev _rx
member. However, the patch did not add the umem to the netdev
structure when user-space passed "best-effort" mode (i.e. select the
fastest possible option available), and zero-copy mode was not
available. This commit fixes that.

Further, this commit makes sure that the umem is not placed into the
netdev structure until the umem is completely setup.

Fixes: c9b47cc1fabc ("xsk: fix bug when trying to use both copy and zero-copy on one queue id")
Signed-off-by: Björn Töpel <bjorn.topel@intel.com>