i40e: proper umem validation, prior enabling AF_XDP ZC
In commit f3fef2b6e1cc ("i40e: Remove umem from VSI") a regression was
introduced. When the VSI was reset, the setup code would try to enable
AF_XDP ZC unconditionally (as long as there was an umem place in the
netdev _rx member). Here, we check that the netdev umem is zero-copy
enabled, prior enabling zero-copy in the driver.

Fixes: f3fef2b6e1cc ("i40e: Remove umem from VSI")
Signed-off-by: Björn Töpel <bjorn.topel@intel.com>