bpf: introduce BPF dispatcher
This builds *very much* on top of Alexei's trampoline ideas. The BPF
dispatcher builds a dispatch table for XDP programs. The table is a
simple binary search model, so lookup is O(log n).

An example: A module/driver allocates a dispatcher. The dispatcher is
shared for all netdevs. Each netdev allocate a slot/id in the
dispatcher and a BPF program. The netdev then uses the dispatcher to
call the correct program with a direct call (actually a tail-call).

Signed-off-by: Björn Töpel <bjorn.topel@intel.com>