xsk: introduce memory type MEM_TYPE_XSK_BUFF_POOL
Here, a new XDP memory type MEM_TYPE_XSK_BUFF_POOL is introduced,
which will supersede the MEM_TYPE_ZERO_COPY.

The new memory type is a single core (single-producer/single-consumer)
XDP buffer pool for AF_XDP UMEMs. To allocate a buffer,
xsk_buff_alloc() is called, and freeing a buffer is done via the
xsk_buff_free() function. To disconnect/disown a buffer from the pool,
xsk_buff_release() is called. This notifies the pool that the freeing
of memory will be done elsewhere, and is not the responsibility of the
pool (analogous to page_pool_release_page()).

The MEM_TYPE_XSK_BUFF_POOL API is defined in net/xdp_sock_buff.h.

Internally there are two pool flavors: aligned and unaligned chunks.

The pool buffer type is struct xdp_buff, and exactly as regular
xdp_buffs, the lifetime of an xdp_buff is restricted to a NAPI
context. In other words, the pool is not replacing xdp_frames.

In addition to introduce the APIs and pool implementations, the AF_XDP
core is migrated to use the new APIs.

Signed-off-by: Björn Töpel <bjorn.topel@intel.com>