i40e, xsk: bulking/staged driver, first step
This is the first step towards a bulking driver. Previously, the i40e
AF_XDP driver had a depth-first way of processing packets: Pick one
packet off the Rx ring, and process it potentially all the way to the
stack. Now, we start of by placing all the completed packets on a
stage, clearing the hardware ring as quickly as possible. When the
ring is cleared, allocate new packets to be received. Finally, the
stage is processed.

Future work will add bulking to the processing step as well.

Rx drop performance up from 19.9 Mpps to 21.4 Mpps.

Signed-off-by: Björn Töpel <bjorn.topel@intel.com>