i40e: refactor i40e_process_skb_fields()
The i40e_process_skb_fields() function is shared by the AF_XDP and
sk_buff path. In upcoming commits, the AF_XDP will move to a bulking
model, and the rx_desc parameter i40e_process_skb_fields() wont be
available. Therefore, the function is restructured to use the qwords
from the writeback descriptor, instead of the descriptor directly.

Signed-off-by: Björn Töpel <bjorn.topel@intel.com>